For 16 - Multiple Product Image : 011 Bootstrap Carousel

+++ WebApplication1/Areas/Admin/Controllers/ProductController.cs
@@ -48,54 +48,65 @@ namespace MyBookWeb.Areas.Admin.Controllers
             else
             {
                 //update
-                productVM.Product = _unitOfWork.Product.Get(u => u.ProductId == id);
+                productVM.Product = _unitOfWork.Product.Get(u => u.ProductId == id, includeProperties: "ProductImages");
                 return View(productVM);
             }
 
         }
         [HttpPost]
-        public IActionResult Upsert(ProductVM productVM, IFormFile? file)
+        public IActionResult Upsert(ProductVM productVM, List<IFormFile> files)
         {
             if (ModelState.IsValid)
             {
-                string wwwRootPath = _webHostEnvironment.WebRootPath;
-                if (file != null)
-                {
-                    string fileName = Guid.NewGuid().ToString() + ".png";
-                    string productPath = Path.Combine(wwwRootPath, @"images\product");
-
-                    if (!string.IsNullOrEmpty(productVM.Product.ImageURL))
-                    {
-                        var oldImagePath = Path.Combine(wwwRootPath, productVM.Product.ImageURL.TrimStart('\\'));
-
-                        if (System.IO.File.Exists(oldImagePath))
-                        {
-                            System.IO.File.Delete(oldImagePath);
-                        }
-                    }
-
-                    using (var fileStream = new FileStream(Path.Combine(productPath, fileName), FileMode.Create))
-                    {
-                        file.CopyTo(fileStream);
-                    }
-
-                    productVM.Product.ImageURL = @"\images\product\" + fileName;
-                }
                 if (productVM.Product.ProductId == 0)
                 {
                     _unitOfWork.Product.Add(productVM.Product);
-                    _unitOfWork.Save();
-                    TempData["success"] = $"{productVM.Product.Title} has been created successfully";
                 }
                 else
                 {
                     _unitOfWork.Product.Update(productVM.Product);
-                    _unitOfWork.Save();
-                    TempData["success"] = $"{productVM.Product.Title} has been updated successfully";
                 }
-                
+                _unitOfWork.Save();
+                string wwwRootPath = _webHostEnvironment.WebRootPath;
+                if (files != null)
+                {
+                    foreach (var file in files)
+                    {
+
+                        string fileName = Guid.NewGuid().ToString() + Path.GetExtension(file.FileName) /*".png"*/;
+                        string productPath = @"images\products\product-" + productVM.Product.ProductId;
+                        string finalPath = Path.Combine(wwwRootPath, productPath);
+                        if (!Directory.Exists(finalPath))
+                        {
+                            Directory.CreateDirectory(finalPath);
+                        }
+
+                        using (var fileStream = new FileStream(Path.Combine(finalPath, fileName), FileMode.Create))
+                        {
+                            file.CopyTo(fileStream);
+                        }
+
+                        ProductImage productImage = new()
+                        {
+                            ImageUrl = @"\" + productPath + @"\" + fileName,
+                            ProductId = productVM.Product.ProductId,
+                        };
+
+                        if (productVM.Product.ProductImages == null)
+                        {
+                            productVM.Product.ProductImages = new List<ProductImage>();
+                        }
+                        productVM.Product.ProductImages.Add(productImage);
+                    }
+                }
+
+                _unitOfWork.Product.Update(productVM.Product);
+                _unitOfWork.Save();
+                TempData["success"] = $"{productVM.Product.Title} has been created/updated successfully";
+
                 return RedirectToAction("Index");
             }
+
             TempData["error"] = $"Failed to add this Product";
             return View();
         }
@@ -118,16 +129,48 @@ namespace MyBookWeb.Areas.Admin.Controllers
                 return Json(new {success = false, message = "Error while deleting"});
             }
 
-            var oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, productToBeDeleted.ImageURL.TrimStart('\\'));
+            string productPath = @"images\products\product-" + id;
+            string finalPath = Path.Combine(_webHostEnvironment.WebRootPath, productPath);
 
-            if (System.IO.File.Exists(oldImagePath))
+            if (Directory.Exists(finalPath))
             {
-                System.IO.File.Delete(oldImagePath);
+                string[] filePaths = Directory.GetFiles(finalPath);
+                foreach (string filePath in filePaths)
+                {
+                    System.IO.File.Delete(filePath);
+                }
+
+                Directory.Delete(finalPath);
             }
+
+
             _unitOfWork.Product.Remove(productToBeDeleted);
             _unitOfWork.Save();
             return Json(new { success = true, message = "Delete successful" });
         } 
+        //[HttpPost]
+        public IActionResult DeleteImage(int imageId)
+        {
+            var imageTobeDeleted = _unitOfWork.ProductImage.Get(u =>u.Id == imageId);
+            int productId = imageTobeDeleted.ProductId;
+            if (imageTobeDeleted != null)
+            {
+                if (!string.IsNullOrEmpty(imageTobeDeleted.ImageUrl))
+                {
+                    var oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, imageTobeDeleted.ImageUrl.TrimStart('\\'));
+
+                    if (System.IO.File.Exists(oldImagePath))
+                    {
+                        System.IO.File.Delete(oldImagePath);
+                    }
+                }
+                _unitOfWork.ProductImage.Remove(imageTobeDeleted);
+                _unitOfWork.Save();
+                TempData["success"] = "The image has been removed successfully";
+            }
+            
+            return RedirectToAction(nameof(Upsert), new { id = productId});
+        }
         #endregion
     }
 }
